+++ src/components/organisms/Header.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { FaPizzaSlice, FaPlus } from 'react-icons/fa';
-import { Logo } from '@/components/atoms/Logo';
-import { ImageButton } from '@/components/atoms/ImageButton';
+import { Logo } from '@/components/molecules/Logo';
 // import './Header.scss';
 
 export const Header = () => {
@@ -12,7 +11,7 @@ export const Header = () => {
         <div className="settings">
           <ul>
             <li className="settings__add">
-              <ImageButton image={FaPlus} />
+              {/* <ImageButton image={FaPlus} /> */}
               {/* <button
                 data-testid="quick-add-task-action"
                 aria-label="Quick and task"
@@ -22,7 +21,7 @@ export const Header = () => {
               </button> */}
             </li>
             <li className="settings__darkmode">
-              <ImageButton image={FaPizzaSlice} />
+              {/* <ImageButton image={FaPizzaSlice} /> */}
               {/* <button
                 data-testid="dark-mode-action"
                 aria-label="Darkmode on/off"
